Basic program to read clipper card ID
This was forked from libfreefare/examples/mifare-desfire-info and uses that basic framework, but specifying the application ID and reading bytes 1-5 in file 8, which is the clipper ID on the back of the card.  This was located using an NFC app on Arlen's phone, because he's cool like that.

This commit is actually from @djcapelis, stealing @arlenarlenarlen's github credentials and computer because reasons.