Basic tool to output tag IDs
This code now reads NFC tags, determines their ID and outputs it.  In
the case of clipper cards, it detects that it is a clipper card and
outputs the clipper ID instead.  I also added a Makefile and updated the
documentation to reflect the new programs.

determine_tag_type.c is a quick and dirty cut out of the code as I was
writing it and it'll output the type of card.  This can be used later to
add new card types that aren't ISO14443-A if for some reason that's a
thing that should be done in the future.

report_tag.c is a quick and dirty cut of the current code, which is
probably useful enough to make as a utility on its own.  guardian.c
currently is the same as this code, but will diverge later when we ask
it to actually do access control.