Implement hold detection
If a tag is held to the reader for more than three seconds, emit an
"h:" to signify that it's been held instead of tapped.  This will
later be used to introduce additional functionality.

These changes also introduced a refactor not only to determine the
time the tag is held against the reader, but to buffer the tag
result, which allows for the RFID tag to be displayed in the syslog
message, along with another few minor odds and ends.  Mostly to
include the tag along with the "h:" event.