Add a watchdog reset error handler
Okay, this code is probably too clever for its own good.  But we have a problem with the doorbot.  libnfc/libfreefare just hangs sometimes.  You make a perfectly reasonable function call.  Just like your mother taught you, polite as can be... and the library just never returns control to your program.  So.  We must take things into our own hands.  We now have a watchdog.  And we turn it on before many of the calls into libnfc/libfreefare.  That watchdog sets a signal for 100ms.  It waits.  If the function returns, the watchdog is disabled.  If not, the watchdog triggers via SIGALRM, longjmps out of the signal handler and starts again from the top.

It's not a great idea.  But it's better than letting a hang prevent us from trying to open the door.